Restructure project and update paths
