+++ experiments/boosted/edl_boosted_experiment_v1.3.py
@@ -1,4 +1,4 @@
-# --------------------- edl_boosted_experiment_v1.0.py ---------------------
+# --------------------- edl_boosted_experiment_v1.3.py ---------------------
 """
 Boosted-Productivity Experiment (bounded version)
 =================================================
