+++ src/ecosim_v1_5.py
@@ -1,4 +1,4 @@
-# ---------------------------- ecosim.py ----------------------------
+# ---------------------------- ecosim_v1_5.py ----------------------------
 """
 Ecosystem-Dominant Logic simulation engine               •  v1.5  (May 2025)
 ====================================================================
